requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
cycler==0.10.0
Cython==0.29.23
Django==3.0
gensim==4.1.2
inflect==5.3.0
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
nltk==3.6.5
numpy==1.21.3
Pillow==8.4.0
pyenchant==3.2.2
pyparsing==3.0.1
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.23
scipy==1.7.1
six==1.16.0
smart-open==5.2.1
sqlparse==0.4.2
tqdm==4.62.3
